[BEAM-6334] Increase timeout (previous was way to small for load tests)

+++ sdks/java/testing/load-tests/src/main/java/org/apache/beam/sdk/loadtests/LoadTestOptions.java
@@ -56,7 +56,7 @@ public interface LoadTestOptions extends PipelineOptions, ApplicationNameOptions
   void setBigQueryTable(String tableName);
 
   @Description("Timeout for a load test expressed in minutes")
-  @Default.Integer(60)
+  @Default.Integer(240)
   Integer getLoadTestTimeout();
 
   void setLoadTestTimeout(Integer timeout);
